requirements.txt
pyserial
libscrc
